+++ schema/buttons.ts
@@ -1,0 +1,34 @@
+export const buttonsSchema = {
+  label: "Buttons",
+  name: "buttons",
+  type: "object",
+  list: true,
+  ui: {
+    component: 'itemListField',
+  },
+  fields: [
+    {
+      label: "Label",
+      name: "label",
+      type: "string",
+    },
+    {
+      label: "Link",
+      name: "link",
+      type: "string",
+    },
+    {
+      type: "string",
+      label: "Button Style",
+      name: "buttonStyle",
+      ui: {
+        component: "buttonControl",
+      },
+    },
+    {
+      label: "Fathom Tracking ID",
+      name: "fathomId",
+      type: "string",
+    },
+  ],
+}