+++ docs/RELEASE_NOTES.md
@@ -1,41 +1,131 @@
-
 # Release Notes
+
+**Version 1.4** | Released 2/17/23
+
+Fixes
+
+- Opengraph image does not show up
+- Section sidebars are not labeled with content data
+- Color Picker gray's show as black
+- Footer not on bottom of page
+- Ornaments make page too wide
+- Page background color not working
+- Upgrade Tina
+
+
+**Version 1.3** | Released 2/8/23
+
+Fixes
+
+- Button font weight
+
+Features
+
+- Theme Preview
+- Sidebar Sections Labeled
+
+**Version 1.2** | Released 1/26/23
+
+Fixes
+
+- Remove all events
+- Restore mobile logo
+- Remove protocol references
+- Remove default posts
+- Allow themes without button or type styles
+- SEO Fix: Revise elements used for feature block headlines
+- Fix layout component props
+- Add license field to package
+
+**Version 1.1** | Released 1/20/23
+
+Features
+
+- Accordian block
+
+Fixes
+
+- Mobile card grid
+- API call loop
+- Background overflow bug
+- Card images are 100% width by default not auto
+- No classes for type or buttons without labels
+
+**Version 1.0**  | Released 1/12/23
+
+Features
+
+- Uses Tina v1
+- All new block settings
+- New theme layer
+- Any number of type and button settings
+- Post and Event collections
+- Fathom analytics added
+
+**Version 0.3.5**  | Released 11/8/22
+
+Features
+
+- Label page sections
+- Feature section mobile styles
+- Sub nav
+- New field control pattern to simplify mobile styles
+- Photo cards border
+- Background transparency
+- First section always has an h1
+- Google analytics updates
+- Global background color for page
+
+Fixes
+
+- Color picker fix
+- Remove draft page concept
+- Desktop nav shift fix
+- Global menu no longer blocks page clicks
+
+**Version 0.3.4**  | Released 6/9/22
+
+Fixes
+
+- Global data fix
+- Tailwind text style fix
 
 **Version 0.3.3**  | Released 6/9/22
 
 Fixes
+
 - Twitter social card
 - Default card data
-
 
 **Version 0.3.2**  | Released 5/27/22
 
 Fixes
-- Update Docs
 
+- Update Docs
 
 **Version 0.3.1**  | Released 5/25/22
 
 Fixes
-- Fix multipage on IPFS
 
+- Fix multipage on IPFS
 
 **Version 0.3.0**  | Released 5/20/22
 
 Features
+
 - Revised Section Settings
 - More global settings
-
 
 **Version 0.2.0**  | Released 4/13/22
 
 Features
-- Upgrade Tina
 
+- Upgrade Tina
 
 **Version 0.1.0**  | Released 12/1/21
 
 Features
+
 - First release.
 - Banner, Feature, Text Cards and Photo Cards
 - Basic Navigation